feacture: changed implement
adding models db, changed logic repository internal sql from ORM, updated config and create utilities

+++ src/services/book_service.py
@@ -1,8 +1,8 @@
 from abc import ABC, abstractmethod
-from fastapi import BackgroundTasks
-from async_timeout import asyncio
-from src.domains import BookFiltered, BookDTO
+from src.domains import BookFiltered, BookDTO, BookEntity
 from src.repositories import IBookRepository, IBookExternalRepository
+from typing import List
+from src.utils import Utils
 
 
 class IBookService(ABC):
@@ -14,17 +14,19 @@ class IBookService(ABC):
     async def delete_book(self, id: str):
         raise NotImplementedError
     
+    @abstractmethod
+    async def save_books_external(self, books: List[BookEntity]):
+        raise NotImplementedError
+    
 
 class BookService(IBookService):
     def __init__(
         self, 
         book_repository: IBookRepository, 
-        book_external_repository: IBookExternalRepository,
-        background_task: BackgroundTasks
+        book_external_repository: IBookExternalRepository
     ):
         self._book_repository = book_repository
         self._book_external_repository = book_external_repository
-        self._background_task = background_task
     
     async def get_books(self, filters: BookFiltered) -> BookDTO:
         books_internal = await self._book_repository.get_books(filters)
@@ -33,10 +35,18 @@ class BookService(IBookService):
         
         books_external = await self._book_external_repository.get_books(filters)
         if books_external.books:
-            # iniciar Background Task para persistir lo obtenido
             return books_external
         
-        return BookDTO(books=[])
+        return BookDTO(books=list(), source=None)
     
     async def delete_book(self, id: str):
+        if not id:
+            return
         await self._book_repository.delete_book(id)
+        
+    async def save_books_external(self, books: List[BookEntity]):
+        print("tarea de persistencia...")
+        chunk_sections = Utils.chunk(books)
+        print(len(chunk_sections))
+        
+
